bug fix for checking for non-existant tiles

+++ create_mosaic_1sec.pro
@@ -49,16 +49,27 @@ PRO create_mosaic_1sec, min_lon, max_lon, min_lat, max_lat, data_type, in_dir, o
 				name_zip=name_matrix[i,j]+'.SRTMGL1.hgt.zip'
 				tmp_name=name_matrix[i,j]+'.hgt'
 			endif
-			
-			spawn, 'unzip '+in_dir+'/'+name_zip
 
-			openr, temp_lun, in_dir+'/'+tmp_name, /get_lun
-			readu, temp_lun, tmp_image
-			free_lun, tmp_lun
-			in_tile_row[*,*,i] = tmp_image
+			full_name_zip = in_dir + '/' + name_zip		
+			full_tmp_name = in_dir + '/' + tmp_name
 
-			;remove uncompressed file.
-			spawn, 'rm '+tmp_name
+			;if file exists, read, if not, fill with 0
+			if (file_test(full_name_zip)) then begin
+
+				spawn, 'unzip '+in_dir+'/'+name_zip
+
+				openr, temp_lun, in_dir+'/'+tmp_name, /get_lun
+				readu, temp_lun, tmp_image
+				free_lun, tmp_lun
+				in_tile_row[*,*,i] = tmp_image
+
+				;remove uncompressed file.
+				spawn, 'rm '+tmp_name
+
+			endif else begin
+				in_tile_row[*,*,i] = 0
+			endelse
+
 		endfor
 
 		;write current row of tiles out to file
